extract importexport settings to own component
