+++ src/pages/Settings.js
@@ -2,33 +2,23 @@
 
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import FileSaver from 'file-saver';
-
-import format from 'date-fns/format';
 
 import Container from 'semantic-ui-react/dist/commonjs/elements/Container';
 import Header from 'semantic-ui-react/dist/commonjs/elements/Header';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Confirm from 'semantic-ui-react/dist/commonjs/addons/Confirm';
 
-import { stateToExport, validData, parseImportData } from '../core/importExport';
+import ImportExport from '../components/Settings/ImportExport';
 
-import { importJSONData, alert, migrateStoreData } from '../actions/app';
 import { truncateTime } from '../actions/time';
 import { truncateProjects } from '../actions/projects';
 
 type SettingsType = {
-  time: any,
-  projects: any,
-  reports: any,
   removeTimeData: () => void,
   removeProjectData: () => void,
-  importData: (data: any) => void,
-  alert: (message: string) => void,
 };
 
 type SettingsState = {
-  confirmImport: boolean,
   confirmRemoveTimesheet: boolean,
   confirmRemoveProjects: boolean,
 }
@@ -43,30 +33,14 @@ class Settings extends Component<SettingsType, SettingsState> {
     this.onRemoveProjects = () => { this.setState({ confirmRemoveProjects: true }); };
     this.onConfirmRemoveProjects = () => this.props.removeProjectData();
 
-    this.onImportData = () => { this.setState({ confirmImport: true }); };
-    this.onOpenImportInput = this.openImportInput.bind(this);
-
     this.onCancelConfirm = () => {
       this.setState({
-        confirmImport: false,
         confirmRemoveTimesheet: false,
         confirmRemoveProjects: false,
       });
     };
 
-    this.onExportData = this.exportData.bind(this);
-
-    // create file upload element
-    const input = document.createElement('input');
-    input.setAttribute('name', 'upload');
-    input.setAttribute('type', 'file');
-
-    input.addEventListener('change', this.handleFileChange.bind(this));
-
-    this.uploadInput = input;
-
     this.state = {
-      confirmImport: false,
       confirmRemoveTimesheet: false,
       confirmRemoveProjects: false,
     };
@@ -76,63 +50,10 @@ class Settings extends Component<SettingsType, SettingsState> {
   onConfirmRemoveTime: () => void;
   onRemoveProjects: () => void;
   onConfirmRemoveProjects: () => void;
-  onExportData: () => void;
-  onOpenImportInput: () => void;
-  onImportData: () => void;
   onCancelConfirm: () => void;
-  uploadInput: HTMLInputElement;
-
-  exportData() {
-    const { time, projects, reports } = this.props;
-
-    const stateToSave = stateToExport({ time, projects, reports });
-
-    const blob = new Blob(
-      [JSON.stringify(stateToSave)],
-      { type: 'application/json;charset=utf-8' },
-    );
-
-    FileSaver.saveAs(blob, `thyme-export_${format(new Date(), 'YYYY-MM-DD')}.json`);
-  }
-
-  openImportInput() {
-    // close confirm modal
-    this.onCancelConfirm();
-
-    // open file dialog
-    this.uploadInput.click();
-  }
-
-  handleImportData(jsonString: string) {
-    try {
-      const importData = parseImportData(JSON.parse(jsonString));
-
-      if (!validData(importData)) {
-        this.props.alert('The provided JSON is not a valid Thyme timesheet');
-        return;
-      }
-
-      this.props.importData(importData);
-
-      this.props.alert('Import successful');
-    } catch (e) {
-      this.props.alert(e.message);
-    }
-  }
-
-  handleFileChange(e) {
-    if (e.target instanceof HTMLInputElement && e.target.files instanceof FileList) {
-      const reader = new FileReader();
-      reader.onload = (ev) => {
-        this.handleImportData(ev.target.result);
-      };
-      reader.readAsText(e.target.files[0]);
-    }
-  }
 
   render() {
     const {
-      confirmImport,
       confirmRemoveTimesheet,
       confirmRemoveProjects,
     } = this.state;
@@ -141,17 +62,7 @@ class Settings extends Component<SettingsType, SettingsState> {
       <Container>
         <Header as="h1">Settings</Header>
 
-        <Header as="h3">Export / Import</Header>
-        <Button color="blue" onClick={this.onExportData}>Export data</Button>
-        <Button color="green" onClick={this.onImportData}>Import data</Button>
-        <Confirm
-          open={confirmImport}
-          content="If you import data it will overwrite your current data. Wish to continue?"
-          confirmButton="Import data"
-          size="mini"
-          onCancel={this.onCancelConfirm}
-          onConfirm={this.onOpenImportInput}
-        />
+        <ImportExport />
 
         <Header as="h3">Delete data</Header>
         <Button color="red" onClick={this.onRemoveTime}>Remove timesheet data</Button>
@@ -184,12 +95,6 @@ class Settings extends Component<SettingsType, SettingsState> {
   }
 }
 
-function mapStateToProps(state) {
-  const { time, projects, reports } = state;
-
-  return { time, projects, reports };
-}
-
 function mapDispatchToProps(dispatch) {
   return {
     removeTimeData() {
@@ -199,16 +104,7 @@ function mapDispatchToProps(dispatch) {
     removeProjectData() {
       dispatch(truncateProjects());
     },
-
-    importData(data) {
-      dispatch(importJSONData(data));
-      dispatch(migrateStoreData());
-    },
-
-    alert(message: string) {
-      dispatch(alert(message));
-    },
   };
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Settings);
+export default connect(null, mapDispatchToProps)(Settings);
